move protocol definitions to new crate
