+++ crates/ott-balancer-bin/Cargo.toml
@@ -19,3 +19,4 @@ tokio.workspace = true
 tokio-tungstenite.workspace = true
 tungstenite.workspace = true
 uuid.workspace = true
+ott-balancer-protocol = { path = "../ott-balancer-protocol" }
