--- crates/ott-balancer-bin/src/protocol/mod.rs
@@ -1,2 +1,0 @@
-pub mod client;
-pub mod monolith;
