+++ crates/ott-balancer-protocol/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "ott-balancer-protocol"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde.workspace = true
+serde_json.workspace = true
+uuid.workspace = true
