+++ crates/ott-balancer-protocol/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod client;
+pub mod monolith;
+pub mod wrappers;
+
+pub use wrappers::*;
